Написал шедулер, но его работу ещё не проверил, так же в энтити микроматча поменял тип начала и конца с локалдейта на локалдейтатайм где-то убрал ненужный код, где то добавил новые методы, начал делать методы для админ панели

+++ src/main/java/ru/inf_fans/web_hockey/service/tournament/TournamentServiceImpl.java
@@ -96,7 +96,7 @@ public class TournamentServiceImpl implements TournamentService {
     @Transactional
     public void removePlayerFromTournament(
             Long tournament_id,
-            int user_id
+            Long user_id
     ) {
         Tournament tournament = tournamentRepository.findTournamentsById(tournament_id);
         if (tournament == null) {
